+++ cli/src/cli_util.rs
@@ -2823,7 +2823,7 @@ fn load_template_aliases(
                 .into());
             }
         };
-        for (decl, item) in table {
+        for (decl, item) in table.iter() {
             let r = item
                 .as_str()
                 .ok_or_else(|| format!("Expected a string, but is {}", item.type_name()))