+++ cli/tests/test_config_command.rs
@@ -94,12 +94,14 @@ fn test_config_list_inline_table() {
     "#,
     );
     let stdout = test_env.jj_cmd_success(test_env.env_root(), &["config", "list", "test-table"]);
-    insta::assert_snapshot!(stdout, @"test-table = { x = true, y = 1 }");
-    // Inner value cannot be addressed by a dotted name path
-    let (stdout, stderr) =
-        test_env.jj_cmd_ok(test_env.env_root(), &["config", "list", "test-table.x"]);
-    insta::assert_snapshot!(stdout, @"");
-    insta::assert_snapshot!(stderr, @"Warning: No matching config key for test-table.x");
+    // Inline tables are expanded
+    insta::assert_snapshot!(stdout, @r"
+    test-table.x = true
+    test-table.y = 1
+    ");
+    // Inner value can also be addressed by a dotted name path
+    let stdout = test_env.jj_cmd_success(test_env.env_root(), &["config", "list", "test-table.x"]);
+    insta::assert_snapshot!(stdout, @"test-table.x = true");
 }
 
 #[test]
@@ -128,8 +130,7 @@ fn test_config_list_array_of_tables() {
         z."key=with whitespace" = []
     "#,
     );
-    // TODO: Perhaps, each value should be listed separately, but there's no
-    // path notation like "test-table[0].x".
+    // Array is a value, so is array of tables
     let stdout = test_env.jj_cmd_success(test_env.env_root(), &["config", "list", "test-table"]);
     insta::assert_snapshot!(stdout, @r###"
     test-table = [{ x = 1 }, { y = ["z"], z = { "key=with whitespace" = [] } }]